Photo service created new method for getting all photos by postID, created method for checking is this the last photo and refactor delete method

+++ src/main/java/com/social/media/service/PhotoService.java
@@ -1,5 +1,6 @@
 package com.social.media.service;
 
+import com.social.media.exception.LastPhotoException;
 import com.social.media.model.entity.Photo;
 import com.social.media.repository.PhotoRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -7,6 +8,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 
@@ -27,11 +29,22 @@ public class PhotoService {
                 new EntityNotFoundException("Photo with id " + id + "not found."));
     }
 
-    public void delete(long id) {
+    public void delete(long postId, long id) {
+        if (isLastPhoto(postId) < 2) {
+            throw new LastPhotoException("In post must be at least one photo!");
+        }
         photoRepository.delete(readById(id));
     }
 
     public Set<Photo> getAll() {
         return new HashSet<>(photoRepository.findAll());
+    }
+
+    public List<Photo> getAllByPost(long postId) {
+        return photoRepository.findAllByPostId(postId);
+    }
+
+    private int isLastPhoto(long postId) {
+        return getAllByPost(postId).size();
     }
 }
